Delete fix

+++ app/Http/Controllers/ReasonComplaintController.php
@@ -87,18 +87,25 @@ class ReasonComplaintController extends Controller
 
     }
 
-    public function destroy(ReasonComplaint $reason_complaint)
+    public function destroy($id)
     {
+       // Eliminar reasoncomplaint
+        $reason_complaint = ReasonComplaint::find($id);
 
-        if ($reason_complaint->delete()) {
-            return response()->json([
-                'message' => 'reason_complaint eliminado correctamente.'
-            ], 204); // 204 No Content = éxito sin datos
-        } else {
-            return response()->json([
-                'error' => 'No se pudo eliminar el reason_complaint.'
-            ], 400); // 400 Bad Request = algo falló en la operación
-        }
+    if (!$reason_complaint) {
+        return response()->json([
+            'error' => 'reasoncomplaint no encontrada.'
+        ], 404);
+    }
 
+    if ($reason_complaint->delete()) {
+        return response()->json([
+            'message' => 'reasoncomplaint eliminada correctamente.'
+        ], 200);
+    }
+
+    return response()->json([
+        'error' => 'No se pudo eliminar la reasoncomplaint.'
+    ], 400);
     }
 }
